src/trusted_datapaper_ds/segmentation/NNU-NET/condaenv.tc44yjzw.requirements.txt
fire
git+https://github.com/julien-blanchon/hiddenlayer.git
git+https://github.com/MIC-DKFZ/acvl_utils.git
git+https://github.com/MIC-DKFZ/dynamic-network-architectures.git
git+https://github.com/Project-MONAI/MONAI#egg=monai
natsort
nibabel
nnunetv2
numpy==1.25.0
open3d==0.16.0
opencv-contrib-python==4.6.0.66
plyfile==0.7.4
pyyaml
scikit-image>=0.19.0
SimpleITK
torch
torchaudio
torchvision
vtk==9.2.6
